// src/app/shared/models/DetallePaciente.ts
export interface DetallePaciente {
    nmid:                 number;
    dsnombres:            string;
    dsapellidos:          string;
    cddocumento:          string;
    dseps:                string;
    dsarl:                string;
    feregistro:           Date;
    febaja:               Date;
    cdusuario:            string;
    dscondicion:          string;
    dsnombresMedicotra:   string;
    dsapellidosMedicotra: string;
    cddocumentoMedicotra: string;
}